Created add note feature to server and removed strict mode from react application

+++ routes/index.js
@@ -1,6 +1,7 @@
 const express = require('express')
 const router = express.Router()
 const Note = require('../models/note')
+const ObjectID = require('mongodb').ObjectID
 
 router.get('/', async (req, res) => {
   try{
@@ -20,29 +21,18 @@ router.get('/', async (req, res) => {
 router.post('/', async (req, res) => {
   console.log('post')
   try{
-    Note.deleteOne({_id: req.body.removed}, err => {
-      if(err) console.log(err);
-      console.log(req.body.removed);
-    })
-    /*let noteList = [];
-    Note.find({})
-    .then(list => {
-      noteList = list.map(note => {
-          return note.id;
-      });
-    })
-    .then(() => {
-      //console.log(data);
-      //const notes = data.notes;
-      //const removed = data.removed;
-      //Update each note
-      /*notes.forEach(element => {
-
+    if(req.body.removed) { //if sent a removed item
+      Note.deleteOne({_id: req.body.removed}, err => {
+        if(err) console.log(err);
       })
-
-
-    })*/
-
+    }else if (req.body.added) { //if added item is sent
+      var newId = new ObjectID();
+      var newNote = new Note({title: req.body.added.title, text: req.body.added.text ,_id: newId})
+      newNote.save(newNote, err => {
+        if(err) console.log(err);
+      });
+      res.json({oldId: req.body.added.id, newId: newId});
+    }
 
   } catch(err) {
     console.log(err)
